refactor: rename+move remaining files and update imports
