+++ packages/compile/src/_shared/subscript.js
@@ -1,8 +1,8 @@
 import util from 'util'
-import R from 'ramda'
 import B from 'bufx'
 import yaml from 'js-yaml'
-import { canonicalName, asort, vlog } from './_shared/helpers.js'
+import R from 'ramda'
+import { canonicalName, asort, vlog } from './helpers.js'
 
 // A subscript is a dimension or an index.
 // Both have the same properties: model name, canonical name, family, values.
