+++ packages/compile/src/generate/code-gen.js
@@ -1,9 +1,11 @@
 import R from 'ramda'
-import ModelLHSReader from './model-lhs-reader.js'
+
+import { asort, lines, strlist, abend, mapIndexed } from '../_shared/helpers.js'
+import { sub, allDimensions, allMappings, subscriptFamilies } from '../_shared/subscript.js'
+import Model from '../model/model.js'
+
 import EquationGen from './equation-gen.js'
-import Model from './Model.js'
-import { sub, allDimensions, allMappings, subscriptFamilies } from './Subscript.js'
-import { asort, lines, strlist, abend, mapIndexed } from './_shared/helpers.js'
+import ModelLHSReader from './model-lhs-reader.js'
 
 export function generateCode(parseTree, opts) {
   return codeGenerator(parseTree, opts).generate()
